+++ src/dog_marker/database/cruds/entry_crud.py
@@ -1,12 +1,11 @@
 import datetime
-from operator import and_, or_
 from typing import Callable, Type
 from uuid import UUID
 
 from result import Result, Err, Ok
 from sqlalchemy import desc, exists
 from sqlalchemy.orm import Session, Query
-from sqlalchemy.sql.operators import is_
+from sqlalchemy.sql.operators import is_, and_, or_
 
 from ..errors import DbNotFoundError
 from ..models import EntryDbModel, CategoryDbModel, EntryImageDbModel, HiddenEntry
@@ -20,13 +19,13 @@ class EntryCRUD:
     def __init__(self, db: Session):
         self.db = db
 
-    def get(self, entry_id: UUID) -> Result[EntryDbModel, Exception]:
+    def get(self, entry_id: UUID):
         result: EntryDbModel | None = self.db.query(EntryDbModel).get(entry_id)
         if result is None:
             return Err(DbNotFoundError(f"Cannot find entry with id {entry_id}"))
         return Ok(result)
 
-    def get_image(self, image_id: int) -> Result[EntryImageDbModel, Exception]:
+    def get_image(self, image_id: int):
         result: EntryImageDbModel | None = self.db.query(EntryImageDbModel).get(image_id)
         if result is None:
             return Err(DbNotFoundError(f"Cannot find entry-image with id {image_id}"))
@@ -35,7 +34,7 @@ class EntryCRUD:
     def query(self) -> Result[Query[Type[EntryDbModel]], Exception]:
         return Ok(self.db.query(EntryDbModel))
 
-    def all(self, page_info: Pagination | None = None) -> Callable[[Query[Type[EntryDbModel]]], list[EntryDbModel]]:
+    def all(self, page_info: Pagination | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> list[EntryDbModel]:
             if page_info:
                 query = query.offset(page_info.skip).limit(page_info.limit)
@@ -44,14 +43,14 @@ class EntryCRUD:
 
         return __internal
 
-    def add(self) -> Callable[[EntryDbModel], EntryDbModel]:
+    def add(self):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             self.db.add(entry)
             return entry
 
         return __internal
 
-    def delete(self, user_id: UUID) -> Callable[[EntryDbModel], EntryDbModel]:
+    def delete(self, user_id: UUID):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             hidden_entry = self.db.query(HiddenEntry).filter_by(entry_id=entry.id, user_id=user_id).first()
             if not hidden_entry:
@@ -65,7 +64,7 @@ class EntryCRUD:
 
         return __internal
 
-    def undo_delete(self, user_id: UUID) -> Callable[[EntryDbModel], EntryDbModel]:
+    def undo_delete(self, user_id: UUID):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             hidden_entry = self.db.query(HiddenEntry).filter_by(entry_id=entry.id, user_id=user_id).first()
             if hidden_entry:
@@ -77,14 +76,14 @@ class EntryCRUD:
 
         return __internal
 
-    def commit(self) -> Callable[[EntryDbModel], EntryDbModel]:
+    def commit(self):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             self.db.commit()
             return entry
 
         return __internal
 
-    def create(self, owner_id: UUID, title: str) -> Result[EntryDbModel, Exception]:
+    def create(self, owner_id: UUID, title: str):
         model = EntryDbModel(user_id=owner_id, title=title)
         return Ok(model)
 
@@ -95,23 +94,21 @@ class EntryCRUD:
 
         return __internal
 
-    def set_description(self, description: str | None = None) -> Callable[[EntryDbModel], EntryDbModel]:
+    def set_description(self, description: str | None = None):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             entry.description = description
             return entry
 
         return __internal
 
-    def set_warning_level(
-        self, warning_level: WarningLevel | warning_levels | None = None
-    ) -> Callable[[EntryDbModel], EntryDbModel]:
+    def set_warning_level(self, warning_level: WarningLevel | warning_levels | None = None):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             entry.warning_level = WarningLevel.from_(warning_level)
             return entry
 
         return __internal
 
-    def set_coordinate(self, longitude: Longitude, latitude: Latitude) -> Callable[[EntryDbModel], EntryDbModel]:
+    def set_coordinate(self, longitude: Longitude, latitude: Latitude):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             entry.longitude = longitude
             entry.latitude = latitude
@@ -119,7 +116,7 @@ class EntryCRUD:
 
         return __internal
 
-    def set_categories(self, categories: list[str] | None) -> Callable[[EntryDbModel], Result[EntryDbModel, Exception]]:
+    def set_categories(self, categories: list[str] | None):
         def __internal(entry: EntryDbModel) -> Result[EntryDbModel, Exception]:
             entry.clear_categories()
             if not categories:
@@ -134,7 +131,7 @@ class EntryCRUD:
 
         return __internal
 
-    def add_image(self, image_path: str | None, image_delete_url: str | None) -> Callable[[EntryDbModel], EntryDbModel]:
+    def add_image(self, image_path: str | None, image_delete_url: str | None):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             if entry.image_delete_url != image_delete_url or entry.image_path != image_path:
                 new_entry_image = EntryImageDbModel(
@@ -146,7 +143,7 @@ class EntryCRUD:
 
         return __internal
 
-    def set_id(self, entry_id: UUID | None) -> Callable[[EntryDbModel], EntryDbModel]:
+    def set_id(self, entry_id: UUID | None):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             if entry_id:
                 entry.id = entry_id
@@ -154,7 +151,7 @@ class EntryCRUD:
 
         return __internal
 
-    def set_create_date(self, create_date: datetime.datetime | None) -> Callable[[EntryDbModel], EntryDbModel]:
+    def set_create_date(self, create_date: datetime.datetime | None):
         def __internal(entry: EntryDbModel) -> EntryDbModel:
             entry.create_date = create_date
             return entry
@@ -169,10 +166,7 @@ class EntryCRUD:
 
         return __internal
 
-    def order_by_coordinate(
-        self,
-        coordinate: Coordinate | None = None,
-    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def order_by_coordinate(self, coordinate: Coordinate | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             if coordinate:
                 query = query.order_by(desc(EntryDbModel.calc_distance(coordinate.longitude, coordinate.latitude)))
@@ -180,9 +174,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_by_date_from(
-        self, date_from: datetime.datetime | None = None
-    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_by_date_from(self, date_from: datetime.datetime | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             if date_from is not None:
                 query = query.filter(EntryDbModel.update_date >= date_from)
@@ -190,9 +182,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_by_warning_level(
-        self, warning_level: WarningLevel | warning_levels | None = None
-    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_by_warning_level(self, warning_level: WarningLevel | warning_levels | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             level_enum = WarningLevel.from_(warning_level)
             query = query.filter(EntryDbModel.warning_level >= level_enum.value)
@@ -200,7 +190,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_by_user(self, user_id: UUID) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_by_user(self, user_id: UUID):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             # noinspection PyTypeChecker
             query = query.filter(EntryDbModel.user_id == user_id)
@@ -208,7 +198,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_show_trash(self, user_id: UUID) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_show_trash(self, user_id: UUID):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             query = query.filter(
                 exists().where(and_(HiddenEntry.user_id == user_id, HiddenEntry.entry_id == EntryDbModel.id))
@@ -217,20 +207,22 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_show_owner_deleted(self) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_show_owner_deleted(self, older_than: datetime.datetime | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
+            and_args = []
+            if older_than is not None:
+                and_args.append(HiddenEntry.update_date <= older_than)
+
             query = query.filter(
                 exists().where(
-                    and_(HiddenEntry.user_id == EntryDbModel.user_id, HiddenEntry.entry_id == EntryDbModel.id)
+                    HiddenEntry.user_id == EntryDbModel.user_id, HiddenEntry.entry_id == EntryDbModel.id, *and_args
                 )
             )
             return query
 
         return __internal
 
-    def filter_owner_deleted(
-        self, ignore_ids: list[UUID] | None = None
-    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_owner_deleted(self, ignore_ids: list[UUID] | None = None):
         ignore_ids = ignore_ids or list()
 
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
@@ -246,9 +238,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_user_deleted(
-        self, user_id: UUID | None = None
-    ) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_user_deleted(self, user_id: UUID | None = None):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             if user_id:
                 query = query.filter(
@@ -258,7 +248,7 @@ class EntryCRUD:
 
         return __internal
 
-    def filter_marked_to_delete(self) -> Callable[[Query[Type[EntryDbModel]]], Query[Type[EntryDbModel]]]:
+    def filter_marked_to_delete(self):
         def __internal(query: Query[Type[EntryDbModel]]) -> Query[Type[EntryDbModel]]:
             query = query.filter(is_(EntryDbModel.mark_to_delete, None))
             return query