+++ changelog.md
@@ -6,6 +6,7 @@
 - Add: Add update update_date if delete or undo delete
 - Add: EntryDbModel.is_deleted
 - Add: Basic BackgroundTask for delete entries
+- Add: DELETE_TRASH_ENTRIES_AFTER_MINUTES, JOB_EXECUTE_INTERVAL_SECONDS, JOB_CLEANUP_INTERVAL_SECONDS
 
 ## v0.4.3
 - Fix: Dependencies for user_entries